1. Changed to type long int:   N and its relatives (neq, Nlocal, Nglobal, etc.);   bandwidth parameters (ml, mu, smu, etc.);   internal indices (i, j, ngroups, etc.) that can have size ~ N;   pivot array, renamed lpivots;   last return flag optional output lsflag; with the following exceptions, where these are of type int:   N, ml, mu, etc. in the C user's call to CVLapackDense and CVLapackBand;   N etc. in the Fortran user's call to FCVLAPACKDENSE and FCVLAPACKBAND   (to be declared as Fortran equivalent of int). (But N, ml, mu arguments in LAPACK user Jacobian fn are still long int.) 2. Interface functions to Blas/Lapack routines declare local int copies of N etc. (intn etc.), and &intn is passed to the BLAS and LAPACK functions in place of &n, etc. 3. Direct linear solver memory block has arrays pivots (int) and lpivots (long int).  File sundials_direct.c has functions NewIntArray and NewLintArray, to create pivots and lpivots, respectively.  LAPACK interfaces use pivots, all other direct solvers use lpivots.

+++ examples/cvodes/parallel/cvsAtmDisp_ASAi_kry_bbd_p.c
@@ -1,7 +1,7 @@
 /*
  * -----------------------------------------------------------------
- * $Revision: 1.1 $
- * $Date: 2007-10-25 20:03:30 $
+ * $Revision: 1.2 $
+ * $Date: 2010-12-01 23:00:48 $
  * -----------------------------------------------------------------
  * Programmer(s): Lukas Jager and Radu Serban @ LLNL
  * -----------------------------------------------------------------
@@ -144,7 +144,7 @@ typedef struct {
   /* Parallel stuff */
   MPI_Comm comm;       /* MPI communicator */
   int myId;            /* process id */ 
-  int npes;            /* total number of processes */
+  long int npes;       /* total number of processes */
   int num_procs[DIM];  /* number of processes in each direction */
   int nbr_left[DIM];   /* MPI ID of "left" neighbor */
   int nbr_right[DIM];  /* MPI ID of "right" neighbor */
@@ -167,7 +167,7 @@ typedef struct {
  */
 
 static int f(realtype t, N_Vector y, N_Vector ydot, void *user_data);
-static int f_local(int Nlocal, realtype t, N_Vector y, 
+static int f_local(long int Nlocal, realtype t, N_Vector y, 
                    N_Vector ydot, void *user_data);
 
 static int fQ(realtype t, N_Vector y, N_Vector qdot, void *user_data);
@@ -175,7 +175,7 @@ static int fQ(realtype t, N_Vector y, N_Vector qdot, void *user_data);
 
 static int fB(realtype t, N_Vector y, N_Vector yB, N_Vector yBdot, 
               void *user_dataB);
-static int fB_local(int NlocalB, realtype t, 
+static int fB_local(long int NlocalB, realtype t, 
                     N_Vector y, N_Vector yB, N_Vector yBdot, 
                     void *user_dataB);
 
@@ -188,10 +188,10 @@ static int fQB(realtype t, N_Vector y, N_Vector yB,
  *------------------------------------------------------------------
  */
 
-static void SetData(ProblemData d, MPI_Comm comm, int npes, int myId,
-                    int *neq, int *l_neq);
+static void SetData(ProblemData d, MPI_Comm comm, long int npes, int myId,
+                    long int *neq, long int *l_neq);
 static void SetSource(ProblemData d);
-static void f_comm(int Nlocal, realtype t, N_Vector y, void *user_data);
+static void f_comm(long int Nlocal, realtype t, N_Vector y, void *user_data);
 static void Load_yext(realtype *src, ProblemData d);
 static void PrintHeader();
 static void PrintFinalStats(void *cvode_mem);
@@ -208,20 +208,20 @@ int main(int argc, char *argv[])
   ProblemData d;
 
   MPI_Comm comm;
-  int npes, npes_needed;
+  long int npes, npes_needed;
   int myId;
  
-  int neq, l_neq;
+  long int neq, l_neq;
 
   void *cvode_mem;
   N_Vector y, q;
   realtype abstol, reltol, abstolQ, reltolQ;
-  int mudq, mldq, mukeep, mlkeep;
+  long int mudq, mldq, mukeep, mlkeep;
 
   int indexB;
   N_Vector yB, qB;
   realtype abstolB, reltolB, abstolQB, reltolQB;
-  int mudqB, mldqB, mukeepB, mlkeepB;
+  long int mudqB, mldqB, mukeepB, mlkeepB;
 
   realtype tret, *qdata, G;
 
@@ -265,7 +265,7 @@ int main(int argc, char *argv[])
   y = N_VNew_Parallel(comm, l_neq, neq);
   N_VConst(ZERO, y);
   
-  /* Allocate and initialize qB (local contributin to cost) */
+  /* Allocate and initialize qB (local contribution to cost) */
   q = N_VNew_Parallel(comm, 1, npes); 
   N_VConst(ZERO, q);
 
@@ -400,8 +400,8 @@ int main(int argc, char *argv[])
  *------------------------------------------------------------------
  */
 
-static void SetData(ProblemData d, MPI_Comm comm, int npes, int myId,
-                    int *neq, int *l_neq)
+static void SetData(ProblemData d, MPI_Comm comm, long int npes, int myId,
+                    long int *neq, long int *l_neq)
 {
   int n[DIM], nd[DIM];
   int dim, size;
@@ -570,7 +570,7 @@ static void SetSource(ProblemData d)
  *------------------------------------------------------------------
  */
 
-static void f_comm(int N_local, realtype t, N_Vector y, void *user_data)
+static void f_comm(long int N_local, realtype t, N_Vector y, void *user_data)
 {
   int id, n[DIM], proc_cond[DIM], nbr[DIM][2];
   ProblemData d;
@@ -683,7 +683,7 @@ static void f_comm(int N_local, realtype t, N_Vector y, void *user_data)
 static int f(realtype t, N_Vector y, N_Vector ydot, void *user_data)
 {
   ProblemData d;
-  int l_neq=1;
+  long int l_neq=1;
   int dim;
 
   d = (ProblemData) user_data;
@@ -698,7 +698,7 @@ static int f(realtype t, N_Vector y, N_Vector ydot, void *user_data)
   return(0);
 }
 
-static int f_local(int Nlocal, realtype t, N_Vector y, 
+static int f_local(long int Nlocal, realtype t, N_Vector y, 
                    N_Vector ydot, void *user_data)
 {
   realtype *Ydata, *dydata, *pdata;
@@ -822,7 +822,7 @@ static int fB(realtype t, N_Vector y, N_Vector yB, N_Vector yBdot,
               void *user_dataB)
 {
   ProblemData d;
-  int l_neq=1;
+  long int l_neq=1;
   int dim;
 
   d = (ProblemData) user_dataB;
@@ -837,7 +837,7 @@ static int fB(realtype t, N_Vector y, N_Vector yB, N_Vector yBdot,
   return(0);
 }
 
-static int fB_local(int NlocalB, realtype t, 
+static int fB_local(long int NlocalB, realtype t, 
                     N_Vector y, N_Vector yB, N_Vector dyB, 
                     void *user_dataB)
 {
